refactor(tests): split CLI tests into per-device files, replace _Impl factories with autospec
Move tests/test_cli.py (1375 lines) to tests/cli/ with per-device-type
files: test_sensor, test_switch, test_light, test_fan, test_number,
test_climate, test_display, test_session, test_smoke.

Replace four hand-rolled _Impl factory functions (_make_device, _make_light,
_make_fan, _make_number) with create_autospec-based typed helpers in
conftest.py. isinstance() checks still work because create_autospec sets
__class__ on the mock. Removes ~200 lines of brittle abstract-method stubs.
