--- tests/test_cli.py
@@ -1,1374 +1,0 @@
-from __future__ import annotations
-
-import importlib
-import json
-import stat
-from enum import Enum, StrEnum
-from io import StringIO
-from pathlib import Path
-from unittest.mock import MagicMock
-
-import pytest
-from rich.console import Console as RichConsole
-from typer.testing import CliRunner
-
-from iaqualink.client import AqualinkAuthState
-from iaqualink.device import (
-    AqualinkBinarySensor,
-    AqualinkClimate,
-    AqualinkDevice,
-    AqualinkFan,
-    AqualinkLight,
-    AqualinkNumber,
-    AqualinkSensor,
-    AqualinkSwitch,
-)
-from iaqualink.exception import AqualinkInvalidParameterException
-from iaqualink.system import SystemStatus, UnsupportedSystem
-
-cli_module = importlib.import_module("iaqualink.cli.app")
-
-app = cli_module.app
-
-
-class FakeSystem:
-    supported = True
-
-    def __init__(self, serial: str, name: str) -> None:
-        self.serial = serial
-        self.name = name
-        self.data = {"device_type": "iaqua"}
-
-    @property
-    def status(self) -> SystemStatus:
-        return SystemStatus.UNKNOWN
-
-    @property
-    def status_translated(self) -> str:
-        return "Unknown"
-
-    async def get_devices(self) -> dict[str, object]:
-        return {}
-
-
-class FakeClient:
-    systems_factory = staticmethod(dict[str, FakeSystem])
-    login_call_count = 0
-
-    def __init__(
-        self,
-        username: str,
-        password: str,
-        event_hooks: dict | None = None,
-    ) -> None:
-        self.username = username
-        self.password = password
-        self._auth_state: AqualinkAuthState | None = None
-
-    @property
-    def auth_state(self) -> AqualinkAuthState | None:
-        return self._auth_state
-
-    @auth_state.setter
-    def auth_state(self, state: AqualinkAuthState | None) -> None:
-        self._auth_state = state
-
-    async def __aenter__(self) -> FakeClient:
-        if self._auth_state is None:
-            await self.login()
-        return self
-
-    async def __aexit__(self, exc_type, exc, tb) -> bool | None:
-        return None
-
-    async def login(self) -> None:
-        type(self).login_call_count += 1
-        self._auth_state = AqualinkAuthState(
-            username=self.username,
-            client_id="session-id",
-            authentication_token="token",
-            user_id="user-id",
-            id_token="id-token",
-            refresh_token="refresh-token",
-        )
-
-    async def get_systems(self) -> dict[str, FakeSystem]:
-        return type(self).systems_factory()
-
-
-@pytest.fixture(autouse=True)
-def reset_fake_client(monkeypatch: pytest.MonkeyPatch) -> None:
-    FakeClient.login_call_count = 0
-    FakeClient.systems_factory = staticmethod(dict[str, FakeSystem])
-    monkeypatch.setattr(cli_module, "AqualinkClient", FakeClient)
-    monkeypatch.setattr(cli_module, "_capture_session", None)
-
-
-def test_capture_flag_creates_file_and_registers_serials(
-    tmp_path: Path,
-) -> None:
-    cookie_jar = tmp_path / "session.json"
-    capture_file = tmp_path / "capture.jsonl"
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"SN001": FakeSystem("SN001", "Pool")}
-    )
-
-    result = CliRunner().invoke(
-        app,
-        [
-            "--capture",
-            str(capture_file),
-            "list-systems",
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert capture_file.exists()
-    session = cli_module._capture_session
-    assert session is not None
-    assert "SN001" in session._literals
-    session.close()
-
-
-def test_list_systems_ignores_malformed_session_jar(tmp_path: Path) -> None:
-    cookie_jar = tmp_path / "session.json"
-    cookie_jar.write_text(json.dumps({"username": "user@example.com"}))
-
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"serial-1": FakeSystem("serial-1", "Backyard")}
-    )
-
-    result = CliRunner().invoke(
-        app,
-        [
-            "list-systems",
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert "Backyard" in result.stdout
-    assert "serial-1" in result.stdout
-    assert "iaqua" in result.stdout
-    assert FakeClient.login_call_count == 1
-
-
-def test_list_systems_writes_owner_only_cookie_jar(tmp_path: Path) -> None:
-    cookie_jar = tmp_path / "session.json"
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"serial-1": FakeSystem("serial-1", "Backyard")}
-    )
-
-    result = CliRunner().invoke(
-        app,
-        [
-            "list-systems",
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert stat.S_IMODE(cookie_jar.stat().st_mode) == 0o600
-
-
-def test_list_devices_reports_ambiguous_system_name(tmp_path: Path) -> None:
-    cookie_jar = tmp_path / "session.json"
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "serial-1": FakeSystem("serial-1", "Backyard"),
-            "serial-2": FakeSystem("serial-2", "Backyard"),
-        }
-    )
-
-    result = CliRunner().invoke(
-        app,
-        [
-            "list-devices",
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--system",
-            "Backyard",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-
-    assert result.exit_code == 1
-    assert "matches multiple systems." in result.stderr
-    assert "Use --system with the serial number instead." in result.stderr
-
-
-def _make_unsupported_system(
-    serial: str = "SN001", name: str = "Pool"
-) -> UnsupportedSystem:
-    data = {"serial_number": serial, "name": name, "device_type": "foo"}
-    return UnsupportedSystem(MagicMock(), data)
-
-
-def test_format_system_line_unsupported() -> None:
-    system = _make_unsupported_system()
-    line = cli_module._format_system_line(system)
-    assert "(unsupported)" in line
-    assert "foo" in line
-
-
-def test_render_device_tree_unsupported_system() -> None:
-    system = _make_unsupported_system(serial="SN001", name="Pool")
-    tree = cli_module._render_device_tree(
-        [("SN001", system)],
-        {"SN001": {}},
-    )
-    buf = StringIO()
-    RichConsole(file=buf, no_color=True, width=120).print(tree)
-    output = buf.getvalue()
-    assert "System type not supported" in output
-    assert "No devices found" not in output
-
-
-def test_list_systems_shows_unsupported_note(tmp_path: Path) -> None:
-    cookie_jar = tmp_path / "session.json"
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"SN001": _make_unsupported_system()}
-    )
-
-    result = CliRunner().invoke(
-        app,
-        [
-            "list-systems",
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert "(unsupported)" in result.output
-
-
-# ---------------------------------------------------------------------------
-# Helpers for unit tests below
-# ---------------------------------------------------------------------------
-
-
-def _make_device(
-    cls: type,
-    label: str = "Dev",
-    value: str | None = "1",
-    value_enum: type[Enum] | None = None,
-) -> AqualinkDevice:
-    """Minimal concrete device of a given base class (bypasses __init__)."""
-    _value_enum = value_enum
-
-    class _Impl(cls):  # type: ignore[valid-type]
-        @property
-        def label(self) -> str:
-            return label
-
-        @property
-        def value(self) -> str | None:
-            return value
-
-        @property
-        def value_enum(self) -> type[Enum] | None:
-            return _value_enum
-
-        @property
-        def value_translated(self) -> str | None:
-            cls_ = _value_enum
-            if cls_ is None or value is None:
-                return None
-            try:
-                return cls_(value).name
-            except ValueError:
-                return None
-
-        @property
-        def name(self) -> str:
-            return label
-
-        @property
-        def manufacturer(self) -> str:
-            return ""
-
-        @property
-        def model(self) -> str:
-            return ""
-
-        @property
-        def is_on(self) -> bool:
-            return bool(value)
-
-        async def turn_on(self) -> None:
-            pass
-
-        async def turn_off(self) -> None:
-            pass
-
-        @property
-        def temperature_unit(self) -> str:
-            return "F"
-
-        @property
-        def current_temperature(self) -> str:
-            return ""
-
-        @property
-        def target_temperature(self) -> str:
-            return ""
-
-        @property
-        def max_temp(self) -> int:
-            return 104
-
-        @property
-        def min_temp(self) -> int:
-            return 40
-
-        async def set_temperature(self, _: int) -> None:
-            pass
-
-        async def _set_temperature(self, _: int) -> None:
-            pass
-
-        @property
-        def brightness_percentage(self) -> int | None:
-            return None
-
-        async def set_brightness_percentage(self, _: int) -> None:
-            pass
-
-        # AqualinkNumber abstract stubs
-        @property
-        def current_value(self) -> float | None:
-            return float(value) if value else None
-
-        @property
-        def min_value(self) -> float:
-            return 0.0
-
-        @property
-        def max_value(self) -> float:
-            return 100.0
-
-        async def _set_value(self, v: float) -> None:
-            pass
-
-    return object.__new__(_Impl)
-
-
-def _render_plain(renderable: object) -> str:
-    buf = StringIO()
-    RichConsole(file=buf, no_color=True, width=120).print(renderable)
-    return buf.getvalue()
-
-
-# Auth state returned by FakeSystemWithAqualink.aqualink — distinct from the
-# state FakeClient writes during _fetch_systems so tests can tell them apart.
-_POST_DEVICE_AUTH = AqualinkAuthState(
-    username="user@example.com",
-    client_id="post-device-session",
-    authentication_token="post-device-token",
-    user_id="user-id",
-    id_token="post-device-id-token",
-    refresh_token="post-device-refresh-token",
-)
-
-
-class FakeSystemWithAqualink(FakeSystem):
-    """FakeSystem that exposes an aqualink client with a known auth state
-    and can return configurable devices from get_devices()."""
-
-    def __init__(
-        self,
-        serial: str,
-        name: str,
-        devices: dict | None = None,
-    ) -> None:
-        super().__init__(serial, name)
-        self._devices = devices or {}
-
-    @property
-    def aqualink(self) -> MagicMock:
-        m = MagicMock()
-        m.auth_state = _POST_DEVICE_AUTH
-        return m
-
-    async def get_devices(self) -> dict:
-        return self._devices
-
-
-# ---------------------------------------------------------------------------
-# _group_devices
-# ---------------------------------------------------------------------------
-
-
-def test_group_devices_all_types() -> None:
-    devices = [
-        ("t1", _make_device(AqualinkClimate, "Heater")),
-        ("l1", _make_device(AqualinkLight, "Light")),
-        ("s1", _make_device(AqualinkSwitch, "Switch")),
-        ("p1", _make_device(AqualinkFan, "Pump")),
-        ("nb1", _make_device(AqualinkNumber, "RPM")),
-        ("b1", _make_device(AqualinkBinarySensor, "Freeze")),
-        ("n1", _make_device(AqualinkSensor, "Temp")),
-    ]
-    groups = cli_module._group_devices(devices)
-    assert [label for _, label, _ in groups] == [
-        "Climate",
-        "Lights",
-        "Switches",
-        "Fans",
-        "Numbers",
-        "Sensors",
-    ]
-    # Binary sensors and regular sensors merged into one group
-    sensors_group = next(g for g in groups if g[1] == "Sensors")
-    assert len(sensors_group[2]) == 2
-
-
-def test_group_devices_climate_not_swallowed_by_switch() -> None:
-    climate = _make_device(AqualinkClimate, "Heater")
-    groups = cli_module._group_devices([("h", climate)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Climate"
-
-
-def test_group_devices_light_not_swallowed_by_switch() -> None:
-    light = _make_device(AqualinkLight, "Spa Light")
-    groups = cli_module._group_devices([("l", light)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Lights"
-
-
-def test_group_devices_fan_grouped_as_fan() -> None:
-    fan = _make_device(AqualinkFan, "Filter Pump")
-    groups = cli_module._group_devices([("p", fan)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Fans"
-
-
-def test_group_devices_number_grouped_as_number() -> None:
-    number = _make_device(AqualinkNumber, "RPM")
-    groups = cli_module._group_devices([("nb", number)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Numbers"
-
-
-def test_group_devices_binary_sensor_in_sensors_group() -> None:
-    # AqualinkBinarySensor has its own explicit _DEVICE_GROUPS entry since it
-    # no longer inherits from AqualinkSensor.
-    binary = _make_device(AqualinkBinarySensor, "Freeze")
-    groups = cli_module._group_devices([("b", binary)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Sensors"
-
-
-def test_group_devices_unknown_type_goes_to_other() -> None:
-    class _Unknown(AqualinkDevice):
-        @property
-        def label(self) -> str:
-            return "X"
-
-        @property
-        def name(self) -> str:
-            return "X"
-
-        @property
-        def manufacturer(self) -> str:
-            return ""
-
-        @property
-        def model(self) -> str:
-            return ""
-
-    device = object.__new__(_Unknown)
-    groups = cli_module._group_devices([("x", device)])
-    assert len(groups) == 1
-    assert groups[0][1] == "Other"
-
-
-def test_group_devices_empty() -> None:
-    assert cli_module._group_devices([]) == []
-
-
-# ---------------------------------------------------------------------------
-# _format_device_line
-# ---------------------------------------------------------------------------
-
-
-def test_format_device_line_sensor_with_value() -> None:
-    device = _make_device(AqualinkSensor, "Pool Temp", "82")
-    text = cli_module._format_device_line("temp", device)
-    assert "Pool Temp" in text.plain
-    assert "[temp]" in text.plain
-    assert "82" in text.plain
-    assert ": " in text.plain
-
-
-def test_format_device_line_sensor_none_value() -> None:
-    device = _make_device(AqualinkSensor, "Pool Temp", None)
-    text = cli_module._format_device_line("temp", device)
-    assert "Pool Temp" in text.plain
-    assert "[temp]" in text.plain
-    assert ": " not in text.plain
-
-
-def test_format_device_line_sensor_empty_value() -> None:
-    device = _make_device(AqualinkSensor, "Pool Temp", "")
-    text = cli_module._format_device_line("temp", device)
-    assert "Pool Temp" in text.plain
-    assert ": " not in text.plain
-
-
-def test_format_device_line_sensor_uses_translated_value() -> None:
-    class _FakeState(StrEnum):
-        OFF = "0"
-        ON = "1"
-
-    device = _make_device(
-        AqualinkSensor, "Pool Temp", "1", value_enum=_FakeState
-    )
-    text = cli_module._format_device_line("temp", device)
-    assert "1" in text.plain
-    assert "(ON)" in text.plain
-
-
-def test_format_device_line_sensor_no_translation_omits_parentheses() -> None:
-    device = _make_device(AqualinkSensor, "Pool Temp", "82")
-    text = cli_module._format_device_line("temp", device)
-    assert "82" in text.plain
-    assert "(" not in text.plain
-
-
-def test_format_device_line_sensor_unknown_enum_value_omits_parentheses() -> (
-    None
-):
-    class _FakeState(StrEnum):
-        OFF = "0"
-        ON = "1"
-
-    device = _make_device(
-        AqualinkSensor, "Pool Temp", "99", value_enum=_FakeState
-    )
-    text = cli_module._format_device_line("temp", device)
-    assert "99" in text.plain
-    assert "(" not in text.plain
-
-
-def test_format_device_line_switch_shows_on_off() -> None:
-    device = _make_device(AqualinkSwitch, "Pool Pump", "1")
-    text = cli_module._format_device_line("pump", device)
-    assert "Pool Pump" in text.plain
-    assert "on" in text.plain
-
-
-def test_format_device_line_light_shows_on_off() -> None:
-    device = _make_device(AqualinkLight, "Spa Light", "1")
-    text = cli_module._format_device_line("spa_light", device)
-    assert "Spa Light" in text.plain
-    assert "on" in text.plain
-
-
-def test_format_device_line_binary_sensor_shows_on_off() -> None:
-    device = _make_device(AqualinkBinarySensor, "Freeze", "1")
-    text = cli_module._format_device_line("freeze", device)
-    assert "Freeze" in text.plain
-    assert "on" in text.plain
-
-
-def test_format_device_line_number_with_unit() -> None:
-    device = _make_number("Filter RPM", "1500", unit="RPM")
-    text = cli_module._format_device_line("rpm", device)
-    assert "Filter RPM" in text.plain
-    assert "1500" in text.plain
-    assert "RPM" in text.plain
-
-
-def test_format_device_line_number_without_unit() -> None:
-    device = _make_number("Pump Speed", "75", unit=None)
-    text = cli_module._format_device_line("speed", device)
-    assert "Pump Speed" in text.plain
-    assert "75" in text.plain
-
-
-def test_format_device_line_number_none_value_shows_dim() -> None:
-    device = _make_number("Pump Speed", None)
-    text = cli_module._format_device_line("speed", device)
-    assert "Pump Speed" in text.plain
-    assert ": " not in text.plain
-
-
-def test_format_device_line_fan_with_preset_shows_preset() -> None:
-    fan = _make_fan(
-        "VSP",
-        "1",
-        supports_presets=True,
-        presets=["LOW", "CUSTOM"],
-        current_preset="CUSTOM",
-    )
-    text = cli_module._format_device_line("vsp", fan)
-    assert "VSP" in text.plain
-    assert "CUSTOM" in text.plain
-
-
-def test_format_device_line_fan_without_preset_no_crash() -> None:
-    fan = _make_fan("VSP", "1", supports_presets=False, supports_turn_on=True)
-    text = cli_module._format_device_line("vsp", fan)
-    assert "VSP" in text.plain
-    assert "on" in text.plain
-
-
-def test_format_device_line_fan_no_controls_shows_dim() -> None:
-    fan = _make_fan("VSP", "1", supports_presets=False)
-    text = cli_module._format_device_line("vsp", fan)
-    assert "VSP" in text.plain
-    assert ": " not in text.plain
-
-
-# ---------------------------------------------------------------------------
-# _render_device_tree — multi-system path
-# ---------------------------------------------------------------------------
-
-
-def test_render_device_tree_single_system_has_systems_root() -> None:
-    system = _make_unsupported_system(serial="SN001", name="Pool")
-    tree = cli_module._render_device_tree([("SN001", system)], {"SN001": {}})
-    output = _render_plain(tree)
-    assert "Systems" in output
-    assert "Pool" in output
-
-
-def test_render_device_tree_multiple_systems() -> None:
-    sys1 = _make_unsupported_system(serial="SN001", name="Pool")
-    sys2 = _make_unsupported_system(serial="SN002", name="Spa")
-    tree = cli_module._render_device_tree(
-        [("SN001", sys1), ("SN002", sys2)],
-        {"SN001": {}, "SN002": {}},
-    )
-    output = _render_plain(tree)
-    assert "Pool" in output
-    assert "Spa" in output
-    assert "Systems" in output
-
-
-def test_render_device_tree_no_systems() -> None:
-    tree = cli_module._render_device_tree([], {})
-    output = _render_plain(tree)
-    assert "No systems found" in output
-
-
-# ---------------------------------------------------------------------------
-# Session jar saved after device / control operations
-# ---------------------------------------------------------------------------
-
-
-def _invoke_with_jar(tmp_path: Path, *args: str) -> tuple:
-    cookie_jar = tmp_path / "session.json"
-    result = CliRunner().invoke(
-        app,
-        [
-            *args,
-            "--username",
-            "user@example.com",
-            "--password",
-            "secret",
-            "--cookie-jar",
-            str(cookie_jar),
-        ],
-    )
-    return result, cookie_jar
-
-
-def test_list_devices_saves_jar_after_device_load(tmp_path: Path) -> None:
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"SN001": FakeSystemWithAqualink("SN001", "Pool")}
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "list-devices")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_status_saves_jar_after_device_load(tmp_path: Path) -> None:
-    FakeClient.systems_factory = staticmethod(
-        lambda: {"SN001": FakeSystemWithAqualink("SN001", "Pool")}
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "status")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_turn_on_saves_jar_after_command(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Pump", "0")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"pump": switch})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "turn-on", "pump")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_set_temperature_saves_jar_after_command(tmp_path: Path) -> None:
-    thermostat = _make_device(AqualinkClimate, "Heater", "72")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink(
-                "SN001", "Pool", {"heater": thermostat}
-            )
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(
-        tmp_path, "set-temperature", "heater", "78"
-    )
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-# ---------------------------------------------------------------------------
-# set-brightness
-# ---------------------------------------------------------------------------
-
-
-def _make_light(
-    label: str = "Light",
-    brightness: int | None = None,
-    effect: str | None = None,
-) -> AqualinkLight:
-    """Concrete AqualinkLight with configurable brightness/effect support."""
-    _brightness = brightness
-    _effect = effect
-    brightness_calls: list[int] = []
-    effect_calls: list[str] = []
-
-    class _Impl(AqualinkLight):
-        @property
-        def label(self) -> str:
-            return label
-
-        @property
-        def name(self) -> str:
-            return label
-
-        @property
-        def manufacturer(self) -> str:
-            return ""
-
-        @property
-        def model(self) -> str:
-            return ""
-
-        @property
-        def is_on(self) -> bool:
-            return True
-
-        async def turn_on(self) -> None:
-            pass
-
-        async def turn_off(self) -> None:
-            pass
-
-        @property
-        def brightness_percentage(self) -> int | None:
-            return _brightness
-
-        async def set_brightness_percentage(self, value: int) -> None:
-            brightness_calls.append(value)
-
-        @property
-        def effect(self) -> str | None:
-            return _effect
-
-        @property
-        def effect_list(self) -> list[str] | None:
-            if _effect is None:
-                return None
-            return [_effect, "Off"]
-
-        async def set_effect(self, name: str) -> None:
-            effect_calls.append(name)
-
-    dev = object.__new__(_Impl)
-    dev._set_brightness_calls = brightness_calls
-    dev._effect_calls = effect_calls
-    return dev
-
-
-def _make_dimmable_light(
-    label: str = "Spa Light", brightness: int = 100
-) -> AqualinkLight:
-    return _make_light(label=label, brightness=brightness)
-
-
-def _make_color_light(
-    label: str = "Color Light", effect: str = "Alpine White"
-) -> AqualinkLight:
-    return _make_light(label=label, effect=effect)
-
-
-def test_set_brightness_succeeds_on_dimmable_light(tmp_path: Path) -> None:
-    light = _make_dimmable_light()
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-brightness", "light", "75")
-    assert result.exit_code == 0
-    assert "75%" in result.stdout
-    assert light._set_brightness_calls == [75]  # type: ignore[attr-defined]
-
-
-def test_set_brightness_fails_on_non_light(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Pump", "0")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"pump": switch})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-brightness", "pump", "50")
-    assert result.exit_code == 1
-    assert "not a light" in result.stderr
-
-
-def test_set_brightness_fails_on_non_dimmable_light(tmp_path: Path) -> None:
-    light = _make_device(AqualinkLight, "Pool Light", "1")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-brightness", "light", "50")
-    assert result.exit_code == 1
-    assert "does not support brightness" in result.stderr
-
-
-def test_set_brightness_saves_jar_after_command(tmp_path: Path) -> None:
-    light = _make_dimmable_light()
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(
-        tmp_path, "set-brightness", "light", "50"
-    )
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-# ---------------------------------------------------------------------------
-# Pump helpers
-# ---------------------------------------------------------------------------
-
-
-def _make_fan(
-    label: str = "VSP",
-    state: str | None = "1",
-    *,
-    supports_turn_on: bool = False,
-    supports_turn_off: bool = False,
-    supports_percentage: bool = False,
-    supports_presets: bool = False,
-    presets: list[str] | None = None,
-    current_preset: str | None = None,
-) -> AqualinkFan:
-    class _Impl(AqualinkFan):
-        @property
-        def label(self) -> str:
-            return label
-
-        @property
-        def name(self) -> str:
-            return label
-
-        @property
-        def manufacturer(self) -> str:
-            return ""
-
-        @property
-        def model(self) -> str:
-            return ""
-
-        @property
-        def supports_turn_on(self) -> bool:
-            return supports_turn_on
-
-        @property
-        def supports_turn_off(self) -> bool:
-            return supports_turn_off
-
-        @property
-        def is_on(self) -> bool:
-            return bool(state)
-
-        async def turn_on(self) -> None:
-            pass
-
-        async def turn_off(self) -> None:
-            pass
-
-        @property
-        def supports_percentage(self) -> bool:
-            return supports_percentage
-
-        async def set_percentage(self, percentage: int) -> None:
-            pass
-
-        @property
-        def supports_presets(self) -> bool:
-            return supports_presets
-
-        @property
-        def preset_modes(self) -> list[str]:
-            return presets or []
-
-        @property
-        def preset_mode(self) -> str | None:
-            return current_preset
-
-        async def set_preset_mode(self, preset_mode: str) -> None:
-            if preset_mode not in (presets or []):
-                raise AqualinkInvalidParameterException(preset_mode)
-
-    return object.__new__(_Impl)
-
-
-def _make_number(
-    label: str = "RPM",
-    state: str | None = "1500",
-    *,
-    min_value: float = 0.0,
-    max_value: float = 3450.0,
-    unit: str | None = "RPM",
-) -> AqualinkNumber:
-    class _Impl(AqualinkNumber):
-        @property
-        def label(self) -> str:
-            return label
-
-        @property
-        def name(self) -> str:
-            return label
-
-        @property
-        def manufacturer(self) -> str:
-            return ""
-
-        @property
-        def model(self) -> str:
-            return ""
-
-        @property
-        def current_value(self) -> float | None:
-            return float(state) if state else None
-
-        @property
-        def min_value(self) -> float:
-            return min_value
-
-        @property
-        def max_value(self) -> float:
-            return max_value
-
-        @property
-        def unit_of_measurement(self) -> str | None:
-            return unit
-
-        async def _set_value(self, value: float) -> None:
-            pass
-
-    return object.__new__(_Impl)
-
-
-# ---------------------------------------------------------------------------
-# turn-on / turn-off with AqualinkFan
-# ---------------------------------------------------------------------------
-
-
-def test_turn_on_pump_with_support(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "0", supports_turn_on=True)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-on", "vsp")
-    assert result.exit_code == 0
-    assert "VSP" in result.output
-
-
-def test_turn_on_pump_without_support_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "0", supports_turn_on=False)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-on", "vsp")
-    assert result.exit_code == 1
-    assert "does not support turn on" in result.stderr
-
-
-def test_turn_off_pump_with_support(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_turn_off=True)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-off", "vsp")
-    assert result.exit_code == 0
-    assert "VSP" in result.output
-
-
-def test_turn_off_pump_without_support_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_turn_off=False)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-off", "vsp")
-    assert result.exit_code == 1
-    assert "does not support turn off" in result.stderr
-
-
-def test_turn_on_sensor_exits(tmp_path: Path) -> None:
-    sensor = _make_device(AqualinkSensor, "Temp", "72")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"temp": sensor})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-on", "temp")
-    assert result.exit_code == 1
-    assert "does not support power controls" in result.stderr
-
-
-def test_turn_off_sensor_exits(tmp_path: Path) -> None:
-    sensor = _make_device(AqualinkSensor, "Temp", "72")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"temp": sensor})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "turn-off", "temp")
-    assert result.exit_code == 1
-    assert "does not support power controls" in result.stderr
-
-
-# ---------------------------------------------------------------------------
-# set-speed
-# ---------------------------------------------------------------------------
-
-
-def test_set_speed_pump_with_support(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_percentage=True)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-speed", "vsp", "75")
-    assert result.exit_code == 0
-    assert "75%" in result.output
-    assert "VSP" in result.output
-
-
-def test_set_speed_pump_without_support_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_percentage=False)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-speed", "vsp", "75")
-    assert result.exit_code == 1
-    assert "does not support speed control" in result.stderr
-
-
-def test_set_speed_non_pump_exits(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Filter", "1")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"filter": switch})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-speed", "filter", "50")
-    assert result.exit_code == 1
-    assert "is not a fan/pump" in result.stderr
-
-
-def test_set_speed_saves_jar(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_percentage=True)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "set-speed", "vsp", "50")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_set_speed_out_of_range_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_percentage=True)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-speed", "vsp", "150")
-    assert result.exit_code == 2
-    assert "150" in result.output
-
-
-# ---------------------------------------------------------------------------
-# set-effect
-# ---------------------------------------------------------------------------
-
-
-def test_set_effect_succeeds_on_color_light(tmp_path: Path) -> None:
-    light = _make_color_light()
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, _ = _invoke_with_jar(
-        tmp_path, "set-effect", "light", "Alpine White"
-    )
-    assert result.exit_code == 0
-    assert "Alpine White" in result.stdout
-    assert light._effect_calls == ["Alpine White"]  # type: ignore[attr-defined]
-
-
-def test_set_effect_fails_on_non_light(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Pump", "0")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"pump": switch})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-effect", "pump", "Alpine White")
-    assert result.exit_code == 1
-    assert "not a light" in result.stderr
-
-
-def test_set_effect_fails_on_non_color_light(tmp_path: Path) -> None:
-    light = _make_device(AqualinkLight, "Pool Light", "1")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, _ = _invoke_with_jar(
-        tmp_path, "set-effect", "light", "Alpine White"
-    )
-    assert result.exit_code == 1
-    assert "does not support color effects" in result.stderr
-
-
-def test_set_effect_saves_jar_after_command(tmp_path: Path) -> None:
-    light = _make_color_light()
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(
-        tmp_path, "set-effect", "light", "Alpine White"
-    )
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_set_effect_rejects_unknown_effect_name(tmp_path: Path) -> None:
-    light = _make_color_light()
-
-    async def _raise(self: AqualinkLight, name: str) -> None:
-        raise AqualinkInvalidParameterException(
-            f"{name!r} isn't a valid effect."
-        )
-
-    type(light).set_effect = _raise  # type: ignore[method-assign]
-
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"light": light})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-effect", "light", "Bogus")
-    assert result.exit_code == 1
-    assert "Bogus" in result.stderr
-
-
-# ---------------------------------------------------------------------------
-# set-preset
-# ---------------------------------------------------------------------------
-
-
-def test_set_preset_pump_with_support(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_presets=True, presets=["Low", "High"])
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-preset", "vsp", "Low")
-    assert result.exit_code == 0
-    assert "Low" in result.output
-    assert "VSP" in result.output
-
-
-def test_set_preset_pump_without_support_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_presets=False)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-preset", "vsp", "Low")
-    assert result.exit_code == 1
-    assert "does not support presets" in result.stderr
-
-
-def test_set_preset_non_pump_exits(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Filter", "1")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"filter": switch})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-preset", "filter", "Low")
-    assert result.exit_code == 1
-    assert "is not a fan/pump" in result.stderr
-
-
-def test_set_preset_saves_jar(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_presets=True, presets=["Low", "High"])
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "set-preset", "vsp", "High")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-def test_set_preset_invalid_preset_exits(tmp_path: Path) -> None:
-    pump = _make_fan("VSP", "1", supports_presets=True, presets=["Low", "High"])
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"vsp": pump})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-preset", "vsp", "Bogus")
-    assert result.exit_code == 1
-    assert "Bogus" in result.stderr
-
-
-# ---------------------------------------------------------------------------
-# set-value
-# ---------------------------------------------------------------------------
-
-
-def test_set_value_number_device(tmp_path: Path) -> None:
-    number = _make_number("RPM", "1500", max_value=3450.0, unit="RPM")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"rpm": number})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-value", "rpm", "2000")
-    assert result.exit_code == 0
-    assert "2000 RPM" in result.output
-
-
-def test_set_value_number_device_without_unit(tmp_path: Path) -> None:
-    number = _make_number("Level", "5", max_value=10.0, unit=None)
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"level": number})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-value", "level", "7")
-    assert result.exit_code == 0
-    assert "7" in result.output
-
-
-def test_set_value_non_number_exits(tmp_path: Path) -> None:
-    switch = _make_device(AqualinkSwitch, "Filter", "1")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"filter": switch})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-value", "filter", "50")
-    assert result.exit_code == 1
-    assert "does not support numeric values" in result.stderr
-
-
-def test_set_value_out_of_range_exits(tmp_path: Path) -> None:
-    number = _make_number("RPM", "1500", max_value=3450.0, unit="RPM")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"rpm": number})
-        }
-    )
-    result, _ = _invoke_with_jar(tmp_path, "set-value", "rpm", "9999")
-    assert result.exit_code == 1
-    assert "out of range" in result.stderr
-
-
-def test_set_value_saves_jar(tmp_path: Path) -> None:
-    number = _make_number("RPM", "1500", max_value=3450.0, unit="RPM")
-    FakeClient.systems_factory = staticmethod(
-        lambda: {
-            "SN001": FakeSystemWithAqualink("SN001", "Pool", {"rpm": number})
-        }
-    )
-    result, cookie_jar = _invoke_with_jar(tmp_path, "set-value", "rpm", "2000")
-    assert result.exit_code == 0
-    data = json.loads(cookie_jar.read_text())
-    assert data["client_id"] == "post-device-session"
-
-
-# ---------------------------------------------------------------------------
-# Smoke imports
-# ---------------------------------------------------------------------------
-
-
-def test_i2d_system_module_importable() -> None:
-    importlib.import_module("iaqualink.systems.i2d.system")
